third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-position-multicol-nested-001-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="target" data-offset-x="11" data-offset-y="-19" data-expected-width="180" data-expected-height="100"></div>\noffsetTop expected -19 but got 63
[FAIL] .target 2
  assert_equals: \n<div class="target" style="background:hotpink;" data-offset-x="173" data-offset-y="-3" data-expected-width="180" data-expected-height="100"></div>\nwidth expected 180 but got 250
Harness: the test ran to completion.
